+++ backend/src/main/java/moheng/runner/warmup/MemberWarmUpRunner.java
@@ -1,6 +1,6 @@
-package moheng.applicationrunner.warmup;
+package moheng.runner.warmup;
 
-import moheng.planner.application.PlannerService;
+import moheng.member.application.MemberService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.ApplicationArguments;
@@ -10,20 +10,22 @@ import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
 @Profile({"dev", "prod"})
-@Order(9)
+@Order(5)
 @Component
-public class PlannerWarmUpRunner implements ApplicationRunner {
-    private static final Logger log = LoggerFactory.getLogger(PlannerWarmUpRunner.class);
-    private final PlannerService plannerService;
+public class MemberWarmUpRunner implements ApplicationRunner {
+    private static final Logger log = LoggerFactory.getLogger(MemberWarmUpRunner.class);
+    private final MemberService memberService;
 
-    public PlannerWarmUpRunner(final PlannerService plannerService) {
-        this.plannerService = plannerService;
+    public MemberWarmUpRunner(final MemberService memberService) {
+        this.memberService = memberService;
     }
+
 
     @Override
     public void run(ApplicationArguments args) throws Exception {
         try {
-            plannerService.findPlannerOrderByRecent(1L);
+            memberService.findByEmail("email");
+            memberService.findMemberAuthorityAndProfileImg(1L);
         } catch (Exception e) {
             log.info("Latency 개선을 위한 JVM Warm Up 처리중입니다.");
         }
