+++ backend/src/main/java/moheng/runner/warmup/PlannerWarmUpRunner.java
@@ -1,6 +1,6 @@
-package moheng.applicationrunner.warmup;
+package moheng.runner.warmup;
 
-import moheng.member.application.MemberService;
+import moheng.planner.application.PlannerService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.ApplicationArguments;
@@ -10,22 +10,20 @@ import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
 @Profile({"dev", "prod"})
-@Order(5)
+@Order(9)
 @Component
-public class MemberWarmUpRunner implements ApplicationRunner {
-    private static final Logger log = LoggerFactory.getLogger(MemberWarmUpRunner.class);
-    private final MemberService memberService;
+public class PlannerWarmUpRunner implements ApplicationRunner {
+    private static final Logger log = LoggerFactory.getLogger(PlannerWarmUpRunner.class);
+    private final PlannerService plannerService;
 
-    public MemberWarmUpRunner(final MemberService memberService) {
-        this.memberService = memberService;
+    public PlannerWarmUpRunner(final PlannerService plannerService) {
+        this.plannerService = plannerService;
     }
-
 
     @Override
     public void run(ApplicationArguments args) throws Exception {
         try {
-            memberService.findByEmail("email");
-            memberService.findMemberAuthorityAndProfileImg(1L);
+            plannerService.findPlannerOrderByRecent(1L);
         } catch (Exception e) {
             log.info("Latency 개선을 위한 JVM Warm Up 처리중입니다.");
         }
